GWZ-23 fix auth issue
add middleware before registering routes
